backend/message-passing/process-agent-message/requirements.txt
google-cloud-firestore==2.9.0
